* renamed `participantportal` package - packages are not camelcase! * moved the ParticipantPortalControllerTest in the same package as the implementation
